% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw.state}
\alias{draw.state}
\title{Shortcut Drawing function}
\usage{
draw.state(
  yvec,
  mpoly,
  breaks = NULL,
  zlim = NULL,
  nbreaks = 11,
  main = "",
  add = FALSE,
  ...
)
}
\description{
Shortcut Drawing function
}
