% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ra_lonlat_to_xy}
\alias{ra_lonlat_to_xy}
\title{Convert global Longitude and Latitude to XY - defaults to NZ(XY)}
\usage{
ra_lonlat_to_xy(lon, lat, crs = 2193)
}
\description{
Convert global Longitude and Latitude to XY - defaults to NZ(XY)
}
