% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ra_xy_to_lonlat}
\alias{ra_xy_to_lonlat}
\title{Convert XY to global Longitude and Latitude - defaults to NZ(XY)}
\usage{
ra_xy_to_lonlat(x, y, crs = 2193)
}
\description{
Convert XY to global Longitude and Latitude - defaults to NZ(XY)
}
